pyproject.toml
[tool.poetry]
name = "acquia-tap-mysql"
version = "0.0.1"
description = "`tap-mysql` is a Singer tap for MySQL, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Josh Lloyd"]
keywords = [
    "ELT",
    "MySQL",
]
license = "Apache 2.0"
packages = [
    { include = "tap_mysql" },
]

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.24.0" }
fs-s3fs = { version = "^1.1.1", optional = true }
pymysql = "^1.0.3"
mysql-replication = "^0.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8 = "^5.0.4"
darglint = "^1.8.1"
black = "^23.1.0"
pyupgrade = "^3.3.1"
mypy = "^1.0.0"
isort = "^5.11.5"
singer-sdk = { version="^0.24.0", extras = ["testing"] }
sqlalchemy-stubs = "^0.4"

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_mysql"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
plugins = "sqlmypy"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-mysql = 'tap_mysql.tap:TapMySQL.cli'
